requirements.txt
pyrogram
aiohttp
aiofiles
python-dotenv
